+++ src/layout/ImageUpload/ImageUpload.module.css
@@ -6,20 +6,6 @@
   align-items: center;
 }
 
-/* Controls Column */
-.controlsContainer {
-  width: 100%;
-  display: flex;
-  flex-direction: column;
-  gap: 1.5rem;
-}
-
-@media (min-width: 1024px) {
-  .controlsContainer {
-    width: 33.333333%;
-  }
-}
-
 /* Canvas Column */
 .canvasContainerWrapper {
   width: 100%;
@@ -66,22 +52,6 @@
   line-height: 1.75rem;
 }
 
-/* Control Sections */
-.controlSection {
-  padding: 1rem;
-  background-color: #f9fafb;
-  border-radius: 0.5rem;
-}
-.controlSection > *:not([hidden]) ~ *:not([hidden]) {
-  margin-top: 1rem;
-}
-
-.label {
-  display: block;
-  font-weight: 500;
-  color: #374151;
-}
-
 /* Icons */
 .icon {
   margin-right: 0.5rem;
@@ -89,118 +59,10 @@
   width: 1.25rem;
 }
 
-/* Buttons */
-.button {
-  flex: 1 1 0%;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  padding: 0.75rem 1rem;
-  font-weight: 600;
-  border-radius: 0.5rem;
-  border: none;
-  cursor: pointer;
-  transition: all 300ms;
-}
-
-.actionButtons,
-.shapeControls {
-  display: flex;
-  gap: 0.5rem;
-  flex-direction: row;
-  align-items: center;
-}
-
-.resetButton {
-  background-color: #e5e7eb;
-  color: #1f2937;
-  box-shadow: 0 1px 2px 0 rgb(0 0 0 / 0.05);
-}
-
-.resetButton:hover {
-  background-color: #d1d5db;
-}
-
-.shapeButton {
-  flex: 1 1 0%;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  padding: 0.5rem;
-  border-radius: 0.375rem;
-  transition:
-    color 150ms ease-in-out,
-    background-color 150ms ease-in-out;
-  background-color: #e5e7eb;
-  color: #374151;
-  border: none;
-  cursor: pointer;
-}
-
-.shapeButton.active {
-  background-color: #2563eb;
-  color: white;
-  box-shadow:
-    0 1px 3px 0 rgb(0 0 0 / 0.1),
-    0 1px 2px -1px rgb(0 0 0 / 0.1);
-}
-
-.cropButton {
-  background-color: #16a34a;
-  color: white;
-  box-shadow:
-    0 4px 6px -1px rgb(0 0 0 / 0.1),
-    0 2px 4px -2px rgb(0 0 0 / 0.1);
-}
-
-.cropButton:hover {
-  background-color: #15803d;
-}
-
-/* Zoom Controls */
-.zoomControls {
-  display: flex;
-  align-items: center;
-  gap: 0.5rem;
-}
-
-.zoomIcon {
-  color: #6b7280;
-}
-
-.zoomSlider {
+.dropZoneWrapper {
   width: 100%;
-  height: 0.5rem;
-  background-color: #e5e7eb;
-  border-radius: 0.5rem;
-  -webkit-appearance: none;
-  appearance: none;
-  cursor: pointer;
-  outline: none;
 }
 
-.zoomSlider::-webkit-slider-thumb {
-  -webkit-appearance: none;
-  appearance: none;
-  width: 1rem;
-  height: 1rem;
-  background: #2563eb;
-  border-radius: 50%;
-  cursor: pointer;
-}
-
-.zoomSlider::-moz-range-thumb {
-  width: 1rem;
-  height: 1rem;
-  background: #2563eb;
-  border-radius: 50%;
-  cursor: pointer;
-  border: none;
-}
-
-.controlsContainer {
-  width: 100%;
-  display: flex;
-  flex-direction: column;
-  gap: 1.5rem;
+.dropZoneWrapper :global(#altinn-drop-zone-image-upload) {
+  border-radius: var(--ds-border-radius-lg);
 }
