--- src/layout/ImageUpload/ImageUploadButton.module.css
@@ -1,7 +1,0 @@
-.dropZoneWrapper :global(#altinn-drop-zone-image-upload) {
-  border-radius: var(--ds-border-radius-lg);
-}
-
-.changeImageLabel {
-  width: fit-content;
-}
